# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/auth:
    post:
      operationId: logAuthEvent
      tags: [sessions]
      summary: Log a safe account security event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthEventInput"
      responses:
        "204":
          description: Logged
  /auth/password-reset-request:
    post:
      operationId: requestPasswordReset
      tags: [sessions]
      summary: Request a password-reset email without revealing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestInput"
      responses:
        "204":
          description: Request accepted
...
components:
  schemas:
...
    AuthEventInput:
      type: object
      required: [event]
      properties:
        event:
          type: string
          enum: [signup, login, logout, password-reset-request, password-changed]
        email:
          type: string
          format: email
        userId:
          type: string
          format: uuid
        device:
          type: string
          maxLength: 160
    PasswordResetRequestInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        device:
          type: string
          maxLength: 160
...