# lib/api-zod/src/generated/types/authEventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthEventInputEvent } from './authEventInputEvent';

export interface AuthEventInput {
  event: AuthEventInputEvent;
  email?: string;
  userId?: string;
  /** @maxLength 160 */
  device?: string;
}